use 64bit timeout values throughout the code, since some timeouts exceeds 4.8 billion microseconds (seconds).
git-svn-id: https://svn.gnumonks.org/trunk/librfid@1426 e0336214-984f-0b4b-a45f-81c69e1f0ede

+++ include/rfid/rfid_layer2.h
@@ -19,7 +19,7 @@ struct rfid_layer2 {
 		int (*transcieve)(struct rfid_layer2_handle *h,
 				  const unsigned char *tx_buf, 
 				  unsigned int tx_len, unsigned char *rx_buf, 
-				  unsigned int *rx_len, unsigned int timeout, 
+				  unsigned int *rx_len, u_int64_t timeout,
 				  unsigned int flags);
 		int (*close)(struct rfid_layer2_handle *h);
 		int (*fini)(struct rfid_layer2_handle *h);
@@ -50,7 +50,7 @@ int rfid_layer2_open(struct rfid_layer2_handle *l2h);
 int rfid_layer2_transcieve(struct rfid_layer2_handle *l2h,
 			   const unsigned char *tx_buf, unsigned int tx_len,
 			   unsigned char *rx_buf, unsigned int *rx_len,
-			   unsigned int timeout, unsigned int flags);
+			   u_int64_t timeout, unsigned int flags);
 int rfid_layer2_close(struct rfid_layer2_handle *l2h);
 int rfid_layer2_fini(struct rfid_layer2_handle *l2h);
